Remove Mutexes
I feel rather silly. When writing the little test program in main.go
I realized that all of my synchronization was happening outside of
the counters themselves. This makes sense since CRDTs are meant
to be used in seperate processes. So the instance on each process
shouldn't have to deal with any concurrency. The syncing is handled
independent of the CRDT code itself. Let's remove the Mutexes
since the local CRDT instances shouldn't dictate the synchronization
scheme. That's handled by other, independent code that can
change based on the usecase.