Add GSet and sets package
A Grow-only set is a set that can only have elements added to it.
We start off the sets package implementing this CRDT. Future plans
include an OR-Set which actually allows for remove of elements from
the set, and optimizations on the OR-Set so that it can avoid
unbounded tombstone growth.

Note the `Set` interface. I've created this so different implementations
of sets can be used interchangeably whenever `Lookup` is the only
operation needed. For example, the `reqInclude` test helper function
should be able to be used for all sets in this package.