Counter Constructors return Pointers
Since all of the recievers for counters are pointers, it makes
sense that the constructors should return pointers as well. This
makes sense since counters shouldn't really have
pass-by-value semantics. This also allows for the types returned
by constructors to implement interfaces directly if we want
to add those in the future (like we did for sets package
which also returns pointers from it's constructors for this
very reason).