Add ORSet
An ORSet is nice because it allows us to remove elements from our set
as well as add elements to our set. The caveat here is that there is
unbounded growth for the adds and removes lists. We'll look into
implementing some way to bound the growth of those lists in upcomming
changes.

Also, I'm defering writing the Serailize and Merge functions for now.