Merge Creates New GSet
I've realized that since I got rid of the mutexes, I think
I can greatly simplify all of these CRDTS. There's no longer
any need for the Serailize functions and the payload structs.
For the GSet, I think Merge should just return a new GSet. This
should make implementing the Merge for the ORSet much easier (I'll
be able to not worry about if one of the GSets is nil).